requirements.txt
Django == 5.2.1
mysqlclient
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
PyJWT==2.9.0
python-dotenv==1.0.1
drf-yasg==1.21.8